Add Prolog rules for Bot-Commit label
Add Prolog rules for the Bot-Commit label, for use by automated bots to
upload changes that are not meant for human review but still need to go
through the commit queue. If Commit-Bot +1 is present, the Code-Review
and Verified labels are removed entirely, allowing for the CL to be
submitted. If the Commit-Bot label is not specified, it is removed
entirely so that it does not show up on normal developers' CLs.
'Commit-Bot' +1 also exempts the change from OWNERS permissions.

TEST=tested in infra/dummies/general-sandbox repository
BUG=chromium:993539

Change-Id: I08190ac8505f7efe03a6a145c9ddb665cc7d4a8d