Add submit requirements for project chromiumos
Rules for creating new submit requirements are as follows:
* We don't create submit requirements for labels that are not blocking.
* For labels that do not define any values, we assume they are
overriding a label definition from the parent project (i.e. to ignore
this label in the child project), hence we create a non-applicable
submit requirement for them.
* For other blocking labels, we create a corresponding submit
requirement for them.

We also convert the blocking label functions to non-blocking, since
submittability rules are now enforced with the newly defined submit
requirements.

Note that for the 'Code-Review' requirement we also set the overrideIf
condition to 'Bot-Commit=+1' since this requirement is currerntly
enforced from All-Projects rules.pl [1]. When this change is submitted,
the Code-Review SR in this change will be authoritative for the project
and needs to behave in the same way as was previously configured.

[1] https://chromium.googlesource.com/All-Projects/+/5ddd22e7537172878280dc2b47373251e842dda7/rules.pl#7

Change-Id: I20a5089282e2d67e3a4a5319c9216edbd8920121