drop Commit-Queue entirely
Sub-settings don't actually inherit, so previous change made the CQ
label unusable in CrOS repos.  Let's drop this entirely to align
the copy behavior across the rest of the host.  The only diff is
changekind:TRIVIAL_REBASE, but since we don't copy the Verified
label in that scenario, copying the CQ label doesn't actually work
since the CQ won't accept it.

Change-Id: Iaedb55624e43420ed42503fb123c30a0a89fa56d